feat(builder): Add components and types for monitoring functionality
Refactored by pulling out the components for monitoring agent flow runs, including AgentFlowList, FlowRunsList, FlowInfo, and FlowRunInfo. These components are integrated into the monitor page, enhancing the development experience.
